+++ hotrod-project/hotrod-torcs/src/main/java/org/hotrod/torcs/rankings/LatestQueriesRanking.java
@@ -9,7 +9,7 @@ import org.hotrod.torcs.QueryExecution;
 
 public class LatestQueriesRanking extends Ranking {
 
-  private static final int DEFAULT_SIZE = 10;
+  private static final int DEFAULT_SIZE = 30;
   private static final int MIN_SIZE = 1;
   private static final int MAX_SIZE = 1000;